코드 정리 v1, DTO contructor 재정의
코드를 보기 좋게 열심히 건드리고 있습니다.

+++ src/main/java/com/anonymous/mentalcare/service/PostService.java
@@ -1,6 +1,10 @@
 package com.anonymous.mentalcare.service;
 
-import com.anonymous.mentalcare.dto.*;
+import com.anonymous.mentalcare.dto.feed.MyPostCommentResponseDto;
+import com.anonymous.mentalcare.dto.feed.MyPostResponseDto;
+import com.anonymous.mentalcare.dto.post.PostDto;
+import com.anonymous.mentalcare.dto.post.RandomPostCommentResponseDto;
+import com.anonymous.mentalcare.dto.post.RandomPostResponseDto;
 import com.anonymous.mentalcare.models.Comment;
 import com.anonymous.mentalcare.models.Post;
 import com.anonymous.mentalcare.models.ReadingPost;
@@ -8,12 +12,14 @@ import com.anonymous.mentalcare.models.User;
 import com.anonymous.mentalcare.repository.PostRepository;
 import com.anonymous.mentalcare.repository.ReadingPostRepository;
 import com.anonymous.mentalcare.security.UserDetailsImpl;
+import com.anonymous.mentalcare.util.ValidateChecker;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
+import javax.validation.Valid;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -24,7 +30,7 @@ public class PostService {
     private final PostRepository postRepository;
     private final ReadingPostRepository readingPostRepository;
 
-    public RandomPostResponseDto getRandomPost(User user) {
+    public RandomPostResponseDto getRandomPost(UserDetailsImpl userDetails) {
         /*
          *  랜덤한 게시글 받아오기
          *  1. 내가 넘기기 하지 않은 글이면서
@@ -40,12 +46,14 @@ public class PostService {
          *  해결방법 : 당장 미봉책으로 size가 0일 땐 일반 검색 하는 식으로 했는데, 멘토님 조언으로 QueryDSL로 변경 예정.
          */
 
+        User user = ValidateChecker.userDetailsIsNull(userDetails);
+
         List<ReadingPost> readingPostList = readingPostRepository.findAllByUser(user);
         List<Long> readingPostIdList = new ArrayList<>();
         for(ReadingPost readingPost : readingPostList){
             readingPostIdList.add(readingPost.getPost().getPostId());
         }
-        System.out.println(readingPostIdList.toString());
+        System.out.println("이미 읽은 글의 Id : " + readingPostIdList);
 
         long qty;
         int idx;
@@ -78,38 +86,37 @@ public class PostService {
             System.out.println("title : " + post.getTitle());
             System.out.println("content : " + post.getContent());
 
-            List<RandomPostCommentResponseDto> commentList = new ArrayList<>();
-            for(Comment comment:post.getCommentList()){
-                commentList.add(new RandomPostCommentResponseDto(comment));
-            }
-            randomPostResponseDto = new RandomPostResponseDto(post, commentList);
+            randomPostResponseDto = new RandomPostResponseDto(post);
             readingPostRepository.save(new ReadingPost(user, post));
         }
 
         return randomPostResponseDto;
     }
 
-    public MyPostResponseDto getMyPost(Long postId, User user){
-        Optional<Post> post = postRepository.findByPostIdAndUser(postId, user);
-        if(!post.isPresent()){
-            throw new NullPointerException("당신의 글이 아니거나 없는 글입니다!");
+    public void savePostService(PostDto.PostWrittenRequestDto postWrittenRequestDto, UserDetailsImpl userDetails){
+        User user = ValidateChecker.userDetailsIsNull(userDetails);
+
+        if(postWrittenRequestDto.getTitle().trim().equals("")){
+            throw new IllegalArgumentException("제목을 입력해주세요.");
+        }else if(postWrittenRequestDto.getContent().trim().equals("")){
+            throw new IllegalArgumentException("내용을 입력해주세요.");
         }
 
-        List<MyPostCommentResponseDto> commentList = new ArrayList<>();
-        for(Comment comment:post.get().getCommentList()){
-            commentList.add(new MyPostCommentResponseDto(comment));
-        }
-
-        return new MyPostResponseDto(post.get(), commentList);
-    }
-
-    public Post savePostService(PostDto.PostWrittenRequestDto postWrittenRequestDto, UserDetailsImpl userDetails){
-        Post post = new Post(postWrittenRequestDto,userDetails.getUser());
+        Post post = new Post(postWrittenRequestDto, user);
         postRepository.save(post);
-        return post;
     }
 
-    public void deletePost(Long postId){
+    public void deletePost(Long postId, UserDetailsImpl userDetails){
+        User user = ValidateChecker.userDetailsIsNull(userDetails);
+
+        Optional<Post> post = postRepository.findById(postId);
+        if(!post.isPresent()){
+            throw new NullPointerException("유효하지 않거나 이미 삭제된 글입니다.");
+        }
+        if(!user.getId().equals(post.get().getUser().getId())){
+            throw new IllegalArgumentException("당신의 게시글이 아닙니다.");
+        }
+
         postRepository.deleteById(postId);
     }
